longevity_rag/requirements.txt

arxiv
langchain
sentence-transformers
chromadb
llama-cpp-python
